src/rai_s2s/pyproject.toml
[tool.poetry]
name = "rai_s2s"
version = "1.0.0"
description = "Speech-to-Speech module for RAI framework"
authors = ["Kajetan Rachwał <kajetan.rachwal@robotec.ai>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
]
packages = [
    { include = "rai_s2s", from = "." },
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10, <3.13"
sounddevice = "^0.4.7"
scipy = "^1.14.0"
pydub = "^0.25.1"
# Optional dependencies (won't be installed unless requested)
elevenlabs = { version = "^1.4.1", optional = true }
openai-whisper = { version = "^20231117", optional = true }
faster-whisper = { version = "^1.1.1", optional = true }
openwakeword = { git = "https://github.com/maciejmajek/openWakeWord.git", branch = "chore/remove-tflite-backend", optional = true }

[tool.poetry.extras]
elevenlabs = ["elevenlabs"]
whisper = ["openai-whisper"]
fasterwhisper = ["faster-whisper"]
wakeword = ["openwakeword"]
all = ["elevenlabs", "openai-whisper", "faster-whisper", "openwakeword"]
